+++ src/frontend/types/libs/altamoon-robust-websocket/index.d.ts
@@ -1,0 +1,20 @@
+import 'altamoon-robust-websocket';
+
+declare module 'altamoon-robust-websocket' {
+  class RobustWebSocket extends WebSocket {
+    constructor(
+      streamUri: string | (() => string | Promise<string>),
+      protocols?: Nullable<string>,
+      options?: {
+        timeout?: number;
+        shouldReconnect?: (event: CloseEvent, ws: RobustWebSocket) => any;
+        automaticOpen?: boolean;
+        ignoreConnectivityEvents?: boolean;
+      },
+    );
+    attempts: number;
+    reconnects: number;
+  }
+
+  export = RobustWebSocket;
+}